+++ blog/database.py
@@ -9,3 +9,11 @@ engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
 sessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 Base = declarative_base()
+
+
+def get_db():
+    db = sessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
